requirements.txt
# ===== Data Science & ML Core =====
pandas
numpy
scikit-learn
xgboost
lightgbm

# ===== Hyperparameter Tuning =====
optuna

# ===== Experiment Tracking =====
clearml

# ===== Web App & Visualization =====
streamlit
streamlit_autorefresh
matplotlib
seaborn

# ===== ONNX Conversion =====
skl2onnx
onnxmltools

# ===== Utilities =====
pyyaml
openpyxl
joblib
